fix : remise de la relation entre réservation et seat a jouren manytomany comme annoncer sur le MCD et mise en place de la requetes pour DTO EventMedium

+++ src/main/java/com/mns/cda/filsrouge/dto/EventLight.java
@@ -2,8 +2,7 @@ package com.mns.cda.filsrouge.dto;
 
 import java.time.LocalDateTime;
 
-public record EventLight(long idEvent,
+public record EventLight(Integer idEvent,
                          String eventName,
                          LocalDateTime eventDate,
-                         String sportName) {
-}
+                         String sportName) {}
